[add][update] 이니시스 결제, tbl_order insert, 주문 페이지 추가

+++ src/main/java/com/forreport/domain/IdPronumVO.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class IdPronumVO {
 
 	private String id;
-	private int pronum;
+	private long pronum;
 	
 }
